refactor modules of kredit to loan_submission, completing loan-submissions endpoints

+++ src/main/java/com/kmkbe/modules/customer/dto/CustomerCompanyDto.java
@@ -34,5 +34,5 @@ public class CustomerCompanyDto {
     private String phone;
     private String ownershipStatus;
     private OffsetDateTime staySince;
-    private BigDecimal stayLength;
+    private Double stayLength;
 }
